requirements.txt
certifi==2019.6.16
cloudinary==1.12.0
Django==2.0.7
django-jet==1.0.8
django-simple-history==2.3.0
djangorestframework==3.8.2
mock==3.0.5
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.1
simplejson==3.16.0
six==1.12.0
urllib3==1.25.3